+++ include/arxml/utilities/model_component_factory.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "arxml/model_elements.hpp"
+#include "arxml/elements.hpp"
 
 namespace arxml::utilities::parser {
