+++ library/model_elements_impl.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include <arxml/model_elements.hpp>
+#include <arxml/elements.hpp>
 
 namespace arxml::model{
